Fix: changing from double to decimal
