+++ lib/core/routes.dart
@@ -1,5 +1,5 @@
 import 'package:warren_everest_challenge/screens/cripto_screen.dart';
 
 var routesDestination = {
-  CriptoScreen.route: (context) => CriptoScreen(),
+  CriptoScreen.route: (context) => const CriptoScreen(),
 };
